backend/requirements.txt
Flask
Flask-Cors
pandas
statsmodels
numpy
scipy
scikit-learn
gunicorn
langchain
langchain-openai
langchain-moonshot
python-dotenv
langchainhub